Improve UI for two level census by closure type
Adds colouring to the `CensusStats` for improved readability. It is now
easier to find the appropriate values across lines.

Omit package and module name from the item list view. The package name
can easily be over 70 characters long, and are thus unreadable on most
monitors any way.
The header panel still shows all the information when hovering over the
relevant list item, and the on-disk results are not affected.

Additionally, homogeneous Arrays do no longer show all entires but show
how many entries of the same closure type they contain.
This reduces memory usage, but also the on-disk size of the two-level
census results.

To make this possible without parsing in various locations, we store structured
information about each Closure Type Census while traversing the heap.
In particular, we store the original `ConstrDesc` and put it into a
single `Text`, where individual entries are split by ";".
This avoids most overhead introduced by the additional data structures.

At long last, we also switch to `Data.Text.IO` for writing the census
results to disk. This seems to remove a memory spike when immediately
writing the results to disk after a census.