Fix Field `closure_desc` in stub.cpp
The `closure_desc` of `InfoProv_` was changed from an `char *` to an
`uint32_t`. This change was mostly done to fix issue:
https://gitlab.haskell.org/ghc/ghc/-/issues/24504

We adapt the stub to these changes by essentially converting the number
stored in `closure_desc` to its string representation, as most tools
expect this format.
For example, the number `10` is turned into the string `"10"`.